# go-server/internal/gen/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stick-It RESP API
  version: 1.0.1
servers:
- url: https://stick-it.lr-projects.de
- url: http://localhost:8081
...
paths:
...
  /api/v2/groups/{groupId}/members:
    delete:
      operationId: deleteMemberFromGroup
      parameters:
      - in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: userId
        in: query
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          description: Deleted
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group or user does not exist
        "409":
          description: User cannot leave as group admin when the group is not empty
      security:
      - token: []
      summary: leave group or delete group when the user is the last group member
      tags:
      - members
    get:
      operationId: getGroupMembers
      parameters:
      - in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/memberResponseDto"
                type: array
          description: Successful response
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group does not exist
      security:
      - token: []
      summary: Get members of a group by ID
      tags:
      - members
    post:
      operationId: joinGroup
      parameters:
      - in: path
        name: groupId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: id of user
        in: query
        name: userId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: inviteUrl to authorize a user to join a private group
        in: query
        name: inviteUrl
        schema:
          $ref: "#/components/schemas/shortString"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groupDto"
          description: Successful response
        "400":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group or user does not exist
        "403":
          description: inviteUrl does not match
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Group or user does not exist
        "409":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: User is already a member
      security:
      - token: []
      summary: Add a member to a group by ID
      tags:
      - members
...
  /api/v2/pins:
    get:
      operationId: getPinImagesByIds
      parameters:
      - description: Comma-separated list of image IDs
        in: query
        name: ids
        schema:
          items:
            $ref: "#/components/schemas/id"
          type: array
      - description: Only pins of this group are returned
        in: query
        name: groupId
        schema:
          $ref: "#/components/schemas/id"
      - description: Only pins of this user are returned
        in: query
        name: userId
        schema:
          $ref: "#/components/schemas/id"
      - description: Describes if the images of the pins should be returned too
        in: query
        name: withImage
        schema:
          default: false
          type: boolean
      - description: Compression level for images (optional)
        in: query
        name: compression
        schema:
          maximum: 100
          minimum: 0
          type: integer
      - description: Height for images (optional)
        in: query
        name: height
        schema:
          maximum: 10000
          minimum: 0
          type: integer
      - description: page number (can only be used when ids is not set). Sorted by
          creation date descending
        in: query
        name: page
        schema:
          type: integer
      - description: page size. Defaults to 20
        in: query
        name: size
        schema:
          default: 20
          type: integer
      - description: only include pins that have been updated after this date.If set
          all deleted pins after this time are returned.
        in: query
        name: updatedAfter
        schema:
          $ref: "#/components/schemas/date"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/pinsSyncDto"
          description: Successful response
      security:
      - token: []
      summary: Get images by IDs
      tags:
      - pins
    post:
      operationId: createPin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/pinRequestDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/pinWithOptionalImageDto"
          description: Successful response
      security:
      - token: []
      summary: Create a new pin
      tags:
      - pins
  /api/v2/pins/{pinId}:
    delete:
      operationId: deletePin
      parameters:
      - in: path
        name: pinId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      responses:
        "200":
          description: successfully deleted
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Pin does not exist
      security:
      - token: []
      summary: Delete a pin by ID
      tags:
      - pins
    get:
      operationId: getPin
      parameters:
      - in: path
        name: pinId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: Describes if the image of the pin should be returned too
        in: query
        name: withImage
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/pinWithOptionalImageDto"
          description: Successful response
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Pin does not exist
      security:
      - token: []
      summary: Get pin information by ID
      tags:
      - pins
  /api/v2/pins/{pinId}/image:
    get:
      operationId: getPinImage
      parameters:
      - in: path
        name: pinId
        required: true
        schema:
          $ref: "#/components/schemas/id"
      - description: "When true, this endpoint redirects directly to the target image\
          \ otherwise the image URL is returned"
        in: query
        name: redirect
        required: false
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            image/*:
              schema:
                $ref: "#/components/schemas/imageUrl"
          description: Successful response
        "301":
          description: Successful redirect to target image
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                $ref: "#/components/schemas/longString"
          description: Pin does not exist
      security:
      - token: []
      summary: Get the image associated with a pin by ID
      tags:
      - pins
...
  /api/v3/sync:
    get:
      operationId: sync
      parameters:
      - description: Syncs created and deleted pins after this date
        in: query
        name: lastSeen
        schema:
          $ref: "#/components/schemas/date"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/syncDto"
          description: Successful response
      security:
      - token: []
      summary: Sync all pins and groups based on last seen date
      tags:
      - pins
components:
  schemas:
    id:
      format: uuid
      type: string
    shortString:
      maxLength: 256
      type: string
    date:
      example: 2017-01-10T14:55:32+01:00
      format: date-time
      type: string
...
    groupDto:
      example:
        lastUpdated: 2017-01-10T14:55:32+01:00
        bestSeason:
          season:
            month: 6
            year: 1
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            seasonNumber: 5
          rank: 2
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          points: 5
        profileImageSmall: https://minio.lr-projects/stick-it/1/1.png
        visibility: 0
        group_admin: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        invite_url: invite_url
        name: name
        link: link
        pinImage: https://minio.lr-projects/stick-it/1/1.png
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        profileImage: https://minio.lr-projects/stick-it/1/1.png
      properties:
        id:
          format: uuid
          type: string
        description:
          maxLength: 256
          type: string
        invite_url:
          maxLength: 256
          type: string
        name:
          maxLength: 256
          type: string
        visibility:
          description: "The visibility of the group. 0 for public, 1 for private"
          type: integer
        group_admin:
          format: uuid
          type: string
        link:
          maxLength: 256
          type: string
        lastUpdated:
          example: 2017-01-10T14:55:32+01:00
          format: date-time
          type: string
        profileImage:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        profileImageSmall:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        pinImage:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        bestSeason:
          $ref: "#/components/schemas/seasonItemDto"
      required:
      - id
      - name
      - visibility
      type: object
    longString:
      maxLength: 500
      type: string
...
    imageUrl:
      example: https://minio.lr-projects/stick-it/1/1.png
      format: url
      type: string
    memberResponseDto:
      example:
        ranking: 0
        profile_image_small: https://minio.lr-projects/stick-it/1/1.png
        userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        username: username
        selectedBatch: 6
      properties:
        userId:
          format: uuid
          type: string
        username:
          maxLength: 256
          type: string
        ranking:
          type: integer
        profile_image_small:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        selectedBatch:
          nullable: true
          type: integer
      required:
      - ranking
      - userId
      - username
      type: object
...
    pinsSyncDto:
      example:
        deleted:
        - null
        - null
        items:
        - image: https://minio.lr-projects/stick-it/1/1.png
          creationUser: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          latitude: -75.5850925717018
          groupId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          creationDate: 2017-01-10T14:55:32+01:00
          longitude: 36.988422590534526
        - image: https://minio.lr-projects/stick-it/1/1.png
          creationUser: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          latitude: -75.5850925717018
          groupId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          creationDate: 2017-01-10T14:55:32+01:00
          longitude: 36.988422590534526
      properties:
        items:
          items:
            $ref: "#/components/schemas/pinWithOptionalImageDto"
          type: array
        deleted:
          items:
            $ref: "#/components/schemas/id"
          type: array
      required:
      - deleted
      - items
      type: object
    pinRequestDto:
      properties:
        image:
          example: iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNk+A8AAQUBAScY42YAAAAASUVORK5CYII=
          format: byte
          type: string
        latitude:
          maximum: 90
          minimum: -90
          type: number
        longitude:
          maximum: 180
          minimum: -180
          type: number
        userId:
          format: uuid
          type: string
        groupId:
          format: uuid
          type: string
        creationDate:
          example: 2017-01-10T14:55:32+01:00
          format: date-time
          type: string
        description:
          maxLength: 1024
          nullable: true
          type: string
      required:
      - groupId
      - image
      - latitude
      - longitude
      - userId
      type: object
    pinWithOptionalImageDto:
      example:
        image: https://minio.lr-projects/stick-it/1/1.png
        creationUser: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        latitude: -75.5850925717018
        groupId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        creationDate: 2017-01-10T14:55:32+01:00
        longitude: 36.988422590534526
      properties:
        id:
          format: uuid
          type: string
        creationDate:
          example: 2017-01-10T14:55:32+01:00
          format: date-time
          type: string
        latitude:
          maximum: 90
          minimum: -90
          type: number
        longitude:
          maximum: 180
          minimum: -180
          type: number
        creationUser:
          format: uuid
          type: string
        image:
          example: https://minio.lr-projects/stick-it/1/1.png
          format: url
          type: string
        groupId:
          format: uuid
          type: string
        description:
          maxLength: 1024
          nullable: true
          type: string
      required:
      - creationDate
      - creationUser
      - groupId
      - id
      - latitude
      - longitude
      type: object
...
    syncDto:
      example:
        groupUpdates:
        - pinsAdded:
          - image: https://minio.lr-projects/stick-it/1/1.png
            creationUser: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            latitude: -75.5850925717018
            groupId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            description: description
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            creationDate: 2017-01-10T14:55:32+01:00
            longitude: 36.988422590534526
          - image: https://minio.lr-projects/stick-it/1/1.png
            creationUser: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            latitude: -75.5850925717018
            groupId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            description: description
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            creationDate: 2017-01-10T14:55:32+01:00
            longitude: 36.988422590534526
          group:
            lastUpdated: 2017-01-10T14:55:32+01:00
            bestSeason:
              season:
                month: 6
                year: 1
                id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
                seasonNumber: 5
              rank: 2
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              points: 5
            profileImageSmall: https://minio.lr-projects/stick-it/1/1.png
            visibility: 0
            group_admin: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            invite_url: invite_url
            name: name
            link: link
            pinImage: https://minio.lr-projects/stick-it/1/1.png
            description: description
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            profileImage: https://minio.lr-projects/stick-it/1/1.png
        - pinsAdded:
          - image: https://minio.lr-projects/stick-it/1/1.png
            creationUser: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            latitude: -75.5850925717018
            groupId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            description: description
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            creationDate: 2017-01-10T14:55:32+01:00
            longitude: 36.988422590534526
          - image: https://minio.lr-projects/stick-it/1/1.png
            creationUser: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            latitude: -75.5850925717018
            groupId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            description: description
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            creationDate: 2017-01-10T14:55:32+01:00
            longitude: 36.988422590534526
          group:
            lastUpdated: 2017-01-10T14:55:32+01:00
            bestSeason:
              season:
                month: 6
                year: 1
                id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
                seasonNumber: 5
              rank: 2
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              points: 5
            profileImageSmall: https://minio.lr-projects/stick-it/1/1.png
            visibility: 0
            group_admin: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            invite_url: invite_url
            name: name
            link: link
            pinImage: https://minio.lr-projects/stick-it/1/1.png
            description: description
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            profileImage: https://minio.lr-projects/stick-it/1/1.png
        deletedPins:
        - null
        - null
      properties:
        deletedPins:
          description: List of ids of deleted pins (unspecific of group)
          items:
            $ref: "#/components/schemas/id"
          type: array
        groupUpdates:
          description: List of groups and their respective pin changes.
          items:
            $ref: "#/components/schemas/syncDto_groupUpdates_inner"
          type: array
      required:
      - deletedPins
      - groupUpdates
      type: object
...
    seasonItemDto:
      example:
        season:
          month: 6
          year: 1
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          seasonNumber: 5
        rank: 2
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        points: 5
      properties:
        id:
          format: uuid
          type: string
        season:
          $ref: "#/components/schemas/seasonDto"
        points:
          description: Number of sticks
          type: integer
        rank:
          description: Rank reached at the end of the season
          type: integer
      required:
      - id
      - points
      - rank
      - season
      type: object
...
    seasonDto:
      example:
        month: 6
        year: 1
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        seasonNumber: 5
      properties:
        id:
          format: uuid
          type: string
        month:
          type: integer
        year:
          type: integer
        seasonNumber:
          type: integer
      required:
      - id
      - month
      - seasonNumber
      - year
      type: object
...
  securitySchemes:
    token:
      bearerFormat: JWT
      scheme: bearer
      type: http