# go-server/internal/gen/server/api_members.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Stick-It RESP API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.1
 */

package genserver

import (
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"
)

// MembersAPIController binds http requests to an api service and writes the service results to the http response
type MembersAPIController struct {
	service      MembersAPIServicer
	errorHandler ErrorHandler
}

// MembersAPIOption for how the controller is set up.
type MembersAPIOption func(*MembersAPIController)

// WithMembersAPIErrorHandler inject ErrorHandler into controller
func WithMembersAPIErrorHandler(h ErrorHandler) MembersAPIOption {
	return func(c *MembersAPIController) {
		c.errorHandler = h
	}
}

// NewMembersAPIController creates a default api controller
func NewMembersAPIController(s MembersAPIServicer, opts ...MembersAPIOption) *MembersAPIController {
	controller := &MembersAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the MembersAPIController
func (c *MembersAPIController) Routes() Routes {
	return Routes{
		"GetGroupMembers": Route{
			"GetGroupMembers",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/members",
			c.GetGroupMembers,
		},
		"JoinGroup": Route{
			"JoinGroup",
			strings.ToUpper("Post"),
			"/api/v2/groups/{groupId}/members",
			c.JoinGroup,
		},
		"DeleteMemberFromGroup": Route{
			"DeleteMemberFromGroup",
			strings.ToUpper("Delete"),
			"/api/v2/groups/{groupId}/members",
			c.DeleteMemberFromGroup,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the MembersAPIController
func (c *MembersAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"GetGroupMembers",
			strings.ToUpper("Get"),
			"/api/v2/groups/{groupId}/members",
			c.GetGroupMembers,
		},
		Route{
			"JoinGroup",
			strings.ToUpper("Post"),
			"/api/v2/groups/{groupId}/members",
			c.JoinGroup,
		},
		Route{
			"DeleteMemberFromGroup",
			strings.ToUpper("Delete"),
			"/api/v2/groups/{groupId}/members",
			c.DeleteMemberFromGroup,
		},
	}
}

// GetGroupMembers - Get members of a group by ID
func (c *MembersAPIController) GetGroupMembers(w http.ResponseWriter, r *http.Request) {
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	result, err := c.service.GetGroupMembers(r.Context(), groupIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// JoinGroup - Add a member to a group by ID
func (c *MembersAPIController) JoinGroup(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	var userIdParam string
	if query.Has("userId") {
		param := query.Get("userId")

		userIdParam = param
	} else {
		c.errorHandler(w, r, &RequiredError{Field: "userId"}, nil)
		return
	}
	var inviteUrlParam string
	if query.Has("inviteUrl") {
		param := query.Get("inviteUrl")

		inviteUrlParam = param
	} else {
	}
	result, err := c.service.JoinGroup(r.Context(), groupIdParam, userIdParam, inviteUrlParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteMemberFromGroup - leave group or delete group when the user is the last group member
func (c *MembersAPIController) DeleteMemberFromGroup(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	groupIdParam := chi.URLParam(r, "groupId")
	if groupIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"groupId"}, nil)
		return
	}
	var userIdParam string
	if query.Has("userId") {
		param := query.Get("userId")

		userIdParam = param
	} else {
		c.errorHandler(w, r, &RequiredError{Field: "userId"}, nil)
		return
	}
	result, err := c.service.DeleteMemberFromGroup(r.Context(), groupIdParam, userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}