# go-server/internal/gen/server/api_pins.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Stick-It RESP API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.1
 */

package genserver

import (
	"encoding/json"
	"net/http"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
)

// PinsAPIController binds http requests to an api service and writes the service results to the http response
type PinsAPIController struct {
	service      PinsAPIServicer
	errorHandler ErrorHandler
}

// PinsAPIOption for how the controller is set up.
type PinsAPIOption func(*PinsAPIController)

// WithPinsAPIErrorHandler inject ErrorHandler into controller
func WithPinsAPIErrorHandler(h ErrorHandler) PinsAPIOption {
	return func(c *PinsAPIController) {
		c.errorHandler = h
	}
}

// NewPinsAPIController creates a default api controller
func NewPinsAPIController(s PinsAPIServicer, opts ...PinsAPIOption) *PinsAPIController {
	controller := &PinsAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the PinsAPIController
func (c *PinsAPIController) Routes() Routes {
	return Routes{
		"GetPinImagesByIds": Route{
			"GetPinImagesByIds",
			strings.ToUpper("Get"),
			"/api/v2/pins",
			c.GetPinImagesByIds,
		},
		"CreatePin": Route{
			"CreatePin",
			strings.ToUpper("Post"),
			"/api/v2/pins",
			c.CreatePin,
		},
		"GetPin": Route{
			"GetPin",
			strings.ToUpper("Get"),
			"/api/v2/pins/{pinId}",
			c.GetPin,
		},
		"DeletePin": Route{
			"DeletePin",
			strings.ToUpper("Delete"),
			"/api/v2/pins/{pinId}",
			c.DeletePin,
		},
		"GetPinImage": Route{
			"GetPinImage",
			strings.ToUpper("Get"),
			"/api/v2/pins/{pinId}/image",
			c.GetPinImage,
		},
		"Sync": Route{
			"Sync",
			strings.ToUpper("Get"),
			"/api/v3/sync",
			c.Sync,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the PinsAPIController
func (c *PinsAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"GetPinImagesByIds",
			strings.ToUpper("Get"),
			"/api/v2/pins",
			c.GetPinImagesByIds,
		},
		Route{
			"CreatePin",
			strings.ToUpper("Post"),
			"/api/v2/pins",
			c.CreatePin,
		},
		Route{
			"GetPin",
			strings.ToUpper("Get"),
			"/api/v2/pins/{pinId}",
			c.GetPin,
		},
		Route{
			"DeletePin",
			strings.ToUpper("Delete"),
			"/api/v2/pins/{pinId}",
			c.DeletePin,
		},
		Route{
			"GetPinImage",
			strings.ToUpper("Get"),
			"/api/v2/pins/{pinId}/image",
			c.GetPinImage,
		},
		Route{
			"Sync",
			strings.ToUpper("Get"),
			"/api/v3/sync",
			c.Sync,
		},
	}
}

// GetPinImagesByIds - Get images by IDs
func (c *PinsAPIController) GetPinImagesByIds(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	var idsParam []string
	if query.Has("ids") {
		idsParam = strings.Split(query.Get("ids"), ",")
	}
	var groupIdParam string
	if query.Has("groupId") {
		param := query.Get("groupId")

		groupIdParam = param
	} else {
	}
	var userIdParam string
	if query.Has("userId") {
		param := query.Get("userId")

		userIdParam = param
	} else {
	}
	var withImageParam bool
	if query.Has("withImage") {
		param, err := parseBoolParameter(
			query.Get("withImage"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "withImage", Err: err}, nil)
			return
		}

		withImageParam = param
	} else {
		var param bool = false
		withImageParam = param
	}
	var compressionParam int32
	if query.Has("compression") {
		param, err := parseNumericParameter[int32](
			query.Get("compression"),
			WithParse[int32](parseInt32),
			WithMinimum[int32](0),
			WithMaximum[int32](100),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "compression", Err: err}, nil)
			return
		}

		compressionParam = param
	} else {
	}
	var heightParam int32
	if query.Has("height") {
		param, err := parseNumericParameter[int32](
			query.Get("height"),
			WithParse[int32](parseInt32),
			WithMinimum[int32](0),
			WithMaximum[int32](10000),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "height", Err: err}, nil)
			return
		}

		heightParam = param
	} else {
	}
	var pageParam int32
	if query.Has("page") {
		param, err := parseNumericParameter[int32](
			query.Get("page"),
			WithParse[int32](parseInt32),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "page", Err: err}, nil)
			return
		}

		pageParam = param
	} else {
	}
	var sizeParam int32
	if query.Has("size") {
		param, err := parseNumericParameter[int32](
			query.Get("size"),
			WithParse[int32](parseInt32),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "size", Err: err}, nil)
			return
		}

		sizeParam = param
	} else {
		var param int32 = 20
		sizeParam = param
	}
	var updatedAfterParam time.Time
	if query.Has("updatedAfter") {
		param, err := parseTime(query.Get("updatedAfter"))
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "updatedAfter", Err: err}, nil)
			return
		}

		updatedAfterParam = param
	} else {
	}
	result, err := c.service.GetPinImagesByIds(r.Context(), idsParam, groupIdParam, userIdParam, withImageParam, compressionParam, heightParam, pageParam, sizeParam, updatedAfterParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// CreatePin - Create a new pin
func (c *PinsAPIController) CreatePin(w http.ResponseWriter, r *http.Request) {
	var pinRequestDtoParam PinRequestDto
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&pinRequestDtoParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertPinRequestDtoRequired(pinRequestDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertPinRequestDtoConstraints(pinRequestDtoParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.CreatePin(r.Context(), pinRequestDtoParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetPin - Get pin information by ID
func (c *PinsAPIController) GetPin(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	pinIdParam := chi.URLParam(r, "pinId")
	if pinIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"pinId"}, nil)
		return
	}
	var withImageParam bool
	if query.Has("withImage") {
		param, err := parseBoolParameter(
			query.Get("withImage"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "withImage", Err: err}, nil)
			return
		}

		withImageParam = param
	} else {
		var param bool = false
		withImageParam = param
	}
	result, err := c.service.GetPin(r.Context(), pinIdParam, withImageParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeletePin - Delete a pin by ID
func (c *PinsAPIController) DeletePin(w http.ResponseWriter, r *http.Request) {
	pinIdParam := chi.URLParam(r, "pinId")
	if pinIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"pinId"}, nil)
		return
	}
	result, err := c.service.DeletePin(r.Context(), pinIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetPinImage - Get the image associated with a pin by ID
func (c *PinsAPIController) GetPinImage(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	pinIdParam := chi.URLParam(r, "pinId")
	if pinIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"pinId"}, nil)
		return
	}
	var redirectParam bool
	if query.Has("redirect") {
		param, err := parseBoolParameter(
			query.Get("redirect"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "redirect", Err: err}, nil)
			return
		}

		redirectParam = param
	} else {
		var param bool = false
		redirectParam = param
	}
	result, err := c.service.GetPinImage(r.Context(), pinIdParam, redirectParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// Sync - Sync all pins and groups based on last seen date
func (c *PinsAPIController) Sync(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	var lastSeenParam time.Time
	if query.Has("lastSeen") {
		param, err := parseTime(query.Get("lastSeen"))
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "lastSeen", Err: err}, nil)
			return
		}

		lastSeenParam = param
	} else {
	}
	result, err := c.service.Sync(r.Context(), lastSeenParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}